+++ src/dreamstate/__tests__/react_tree_sync.test.ts
@@ -4,7 +4,7 @@ import { act } from "react-dom/test-utils";
 
 import { useManager } from "@/dreamstate";
 import { createProvider } from "@/dreamstate/core/provision/createProvider";
-import { getCurrent } from "@/dreamstate/core/registry/getCurrent";
+import { getCurrent } from "@/dreamstate/test-utils/registry/getCurrent";
 import { TestContextManager } from "@/fixtures";
 
 describe("React tree for provided and consumed components", () => {