+++ src/dreamstate/core/consumption/useManager.test.ts
@@ -4,7 +4,7 @@ import { act } from "react-dom/test-utils";
 
 import { createProvider, useManager } from "@/dreamstate";
 import { CONTEXT_SERVICES_REGISTRY } from "@/dreamstate/core/internals";
-import { getCurrent } from "@/dreamstate/core/registry/getCurrent";
+import { getCurrent } from "@/dreamstate/test-utils/registry/getCurrent";
 import { registerService } from "@/dreamstate/test-utils/registry/registerService";
 import { nextAsyncQueue } from "@/dreamstate/test-utils/utils/nextAsyncQueue";
 import { TAnyContextManagerConstructor, TAnyObject, TCallable } from "@/dreamstate/types";